% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_distance_features}
\alias{create_distance_features}
\title{Create distance features}
\usage{
create_distance_features(how_far_df)
}
\arguments{
\item{how_far_df}{the output of how_far_is_it()}
}
\value{
a tibble
}
\description{
Create distance features
}
\examples{

address <- "Amager Boulevard 126, 2300 København S"
token <- "eyJkcGZ4IjogImh2b3JsYW5ndGVyZGVyIiwgInByaXZzIjogInIxWjByMEYwazZCdFdxUWNPVXlrQi95NlNVcEp2MlFiZ3lYZXRxNEhZNFhPLzNZclcwK0s5dz09In0.fP4JWis69HmaSg5jVHiK8nemiCu6VaMULSGGJyK4D4PkWq4iA1+nSHWMaHxepKwJ83sEiy9nMNZhv7BcktRNrA"

how_far_df <- how_far_is_it(address, token)
how_far_features <- create_distance_features(how_far_df)
}
