% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{how_far_is_it}
\alias{how_far_is_it}
\title{Get distance to POIs}
\usage{
how_far_is_it(address, token)
}
\arguments{
\item{address}{the address to get POIs for}

\item{token}{the token for the API}
}
\value{
a tibble
}
\description{
This function get s distances to POIs based on an address
}
\examples{

address <- "Amager Boulevard 126, 2300 København S"
token <- "eyJkcGZ4IjogImh2b3JsYW5ndGVyZGVyIiwgInByaXZzIjogInIxWjByMEYwazZCdFdxUWNPVXlrQi95NlNVcEp2MlFiZ3lYZXRxNEhZNFhPLzNZclcwK0s5dz09In0.fP4JWis69HmaSg5jVHiK8nemiCu6VaMULSGGJyK4D4PkWq4iA1+nSHWMaHxepKwJ83sEiy9nMNZhv7BcktRNrA"

how_far_is_it(address, token)
}
